\name{coombs-package}
\alias{coombs-package}
\alias{coombs}
\docType{package}
\title{\packageTitle{coombs}}
\description{\packageDescription{coombs}}
\section{Package Content}{\packageIndices{coombs}}
\author{\packageAuthor{coombs}}
\section{Maintainer}{\packageMaintainer{coombs}}
\keyword{package}
