\name{DCC}
\alias{DCC}
\title{Double Cancellation Check}
\description{
This function should only be used internally. It counts all the intances of the double cancellation axiom for all the 3 X 3 and 4 X 4 matrices in the data.
}
\usage{
DCC(PDC, TbTt, Pairs, Rows, Cols)
}
\arguments{
   \item{PDC}{XXXX.}
   \item{TbTt}{XXXX.}
   \item{Pairs}{XXXX.}
   \item{Rows}{XXXX.}
   \item{Cols}{XXXX.}
}
\value{
A vector of valid countings of the double cancellation axiom.
}
\examples{
##### USED INTERNALLY #####
}
