\name{four_tuples}
\alias{four_tuples}
\title{Get all four tuples of variables}
\description{
This function should only be used internally. Lists all the four tuples of varibles, representing the possible midpoints in the data.
}
\usage{
four_tuples(X)
}
\arguments{
   \item{X}{XXXX.}
}
\value{
A vector of all the four tuples of variables.
}
\examples{
##### USED INTERNALLY #####
}
