% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{plots}
\alias{plots}
\alias{plot.CS_ranking}
\title{S3Methods for Ploting}
\usage{
\method{plot}{CS_ranking}(x, ...)

}
\arguments{
\item{x}{Object from \code{CS_ranking} package.}

\item{...}{Additional arguments}
}
\value{
Plots \code{CS_ranking} object
}
\description{
Plots for \code{CS_ranking} objects
}
\author{
Vithor Rosa Franco <vithorfranco@gmail.com>
}