% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{prints}
\alias{prints}
\alias{print.CS_ranking}
\title{S3Methods for Printing}
\usage{
\method{print}{CS_ranking}(x, ...)

}
\arguments{
\item{x}{Object from \code{CS_ranking} package.}

\item{...}{Additional arguments}
}
\value{
Prints \code{CS_ranking} object
}
\description{
Prints for \code{CS_ranking} objects
}
\author{
Vithor Rosa Franco <vithorfranco@gmail.com>
}