% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarys.R
\name{summarys}
\alias{summarys}
\alias{summary.CS_ranking}
\title{S3Methods for summarying}
\usage{
\method{summary}{CS_ranking}(object, ...)

}
\arguments{
\item{object}{An \code{CS_ranking} object.}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
Summarise \code{CS_ranking} object
}
\description{
Summaries for \code{CS_ranking} objects
}
\author{
Vithor Rosa Franco <vithorfranco@gmail.com>
}