\name{coombs-package}
\alias{coombs-package}
\alias{coombs}
\docType{package}
\title{
\packageTitle{coombs}
}
\description{
\packageDescription{coombs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{coombs}
\packageIndices{coombs}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{coombs}

Maintainer: \packageMaintainer{coombs}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
